Move AuthContextResolver to package resolver
