+++ cmd/auth/main.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/skygeario/skygear-server/pkg/core/auth/authinfo"
-	defaultAuthn "github.com/skygeario/skygear-server/pkg/core/auth/authn/default"
+	"github.com/skygeario/skygear-server/pkg/core/auth/authn/resolver"
 	"github.com/skygeario/skygear-server/pkg/core/auth/authtoken"
 
 	"github.com/kelseyhightower/envconfig"
@@ -53,7 +53,7 @@ func main() {
 	}
 
 	srv.SetAuthContextResolverFactory(
-		defaultAuthn.AuthContextResolverFactory{ProviderGraph: authDependency},
+		resolver.AuthContextResolverFactory{ProviderGraph: authDependency},
 	)
 
 	handler.AttachSignupHandler(&srv, authDependency)
